medgemma_challenge/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.5
pydantic-settings==2.11.0
python-dotenv==1.0.1
requests==2.32.5
pytest==8.4.2

# Optional for local MedGemma inference:
transformers>=4.57.1
torch>=2.2.0
